# sqlstats: fix race condition in ss_mem_iterator