@kyle-a-wong 2025-01-17
Bench diff results when running:
`benchdiff  ./pkg/sql/sqlstats/persistedsqlstats -b -r 'BenchmarkSqlStatsPersisted'`
```

name                                                                                       old time/op    new time/op    delta
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12     120ms ±34%     104ms ± 5%  -13.04%  (p=0.028 n=10+9)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                  18.8s ±18%     18.3s ±19%     ~     (p=0.579 n=10+10)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12       103ms ±14%     111ms ±33%     ~     (p=0.400 n=9+10)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12    80.4ms ±20%    76.3ms ±13%     ~     (p=0.661 n=10+9)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                  18.0s ± 8%     19.1s ±32%     ~     (p=0.673 n=8+9)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12       113ms ±11%     116ms ±13%     ~     (p=0.549 n=9+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12    81.0ms ± 6%    83.0ms ±13%     ~     (p=0.573 n=8+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                  17.1s ±17%     17.8s ±17%     ~     (p=0.278 n=9+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12       125ms ±10%     119ms ±12%     ~     (p=0.063 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12    88.7ms ± 5%    85.9ms ± 7%     ~     (p=0.053 n=9+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                  16.9s ±14%     16.4s ± 3%     ~     (p=0.897 n=10+8)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       124ms ±21%     125ms ±22%     ~     (p=1.000 n=10+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12    97.0ms ±21%    89.4ms ± 7%     ~     (p=0.113 n=10+9)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                  17.3s ±27%     16.3s ± 8%     ~     (p=0.604 n=10+9)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12       145ms ±19%     146ms ±17%     ~     (p=0.684 n=10+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                  16.8s ±27%     15.5s ± 4%     ~     (p=0.200 n=9+8)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       179ms ± 9%     186ms ±13%     ~     (p=0.321 n=8+9)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     136ms ±11%     142ms ±10%     ~     (p=0.222 n=9+9)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                  15.5s ±14%     15.7s ±13%     ~     (p=0.315 n=9+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12       229ms ±40%     203ms ±17%     ~     (p=0.182 n=10+9)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12     166ms ±17%     170ms ±16%     ~     (p=0.436 n=10+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                  6.78s ±21%     6.52s ±24%     ~     (p=0.211 n=9+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12      6.79ms ±63%    6.11ms ±13%     ~     (p=0.604 n=10+9)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12    4.55ms ± 7%    4.45ms ±10%     ~     (p=0.321 n=8+9)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                  6.50s ±32%     6.21s ±14%     ~     (p=0.842 n=9+10)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12      6.14ms ±24%    6.07ms ±13%     ~     (p=0.905 n=9+10)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12    4.63ms ± 4%    4.63ms ± 9%     ~     (p=0.666 n=9+9)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                  6.24s ±25%     6.59s ±23%     ~     (p=0.549 n=9+10)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12      6.05ms ± 9%    6.35ms ±13%     ~     (p=0.280 n=10+10)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12    4.64ms ± 4%    5.04ms ±21%     ~     (p=0.315 n=8+10)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                  18.7s ±10%     19.0s ±26%     ~     (p=0.815 n=8+9)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       101ms ±25%      95ms ±15%     ~     (p=0.447 n=10+9)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12    72.8ms ± 9%    81.8ms ±32%     ~     (p=0.083 n=8+10)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                  6.39s ±35%     6.82s ±15%     ~     (p=0.156 n=9+10)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12      7.21ms ±54%    6.10ms ±23%     ~     (p=0.278 n=10+9)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12    4.73ms ±15%    4.60ms ±20%     ~     (p=0.360 n=10+8)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                  17.5s ±21%     17.7s ±19%     ~     (p=0.912 n=10+10)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       140ms ± 9%     145ms ±40%     ~     (p=0.743 n=8+9)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     103ms ±13%     106ms ±21%     ~     (p=0.905 n=9+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                  6.88s ± 8%     7.21s ±12%     ~     (p=0.130 n=8+8)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12      5.37ms ± 5%    5.56ms ±18%     ~     (p=0.549 n=9+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12    4.29ms ±11%    4.55ms ±25%     ~     (p=0.393 n=10+10)

name                                                                                       old alloc/op   new alloc/op   delta
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                 19.3GB ± 2%    19.0GB ± 1%   -1.81%  (p=0.009 n=10+8)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                 18.0GB ± 1%    18.0GB ± 1%     ~     (p=0.631 n=10+10)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12       522MB ± 1%     517MB ± 1%     ~     (p=0.053 n=10+9)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12     320MB ± 1%     319MB ± 1%     ~     (p=0.481 n=10+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                 18.0GB ± 1%    18.0GB ± 1%     ~     (p=0.739 n=10+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12     368MB ± 1%     370MB ± 1%     ~     (p=0.089 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12       637MB ± 2%     637MB ± 1%     ~     (p=0.971 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12     409MB ± 2%     410MB ± 1%     ~     (p=0.481 n=10+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                 16.4GB ± 0%    16.4GB ± 0%     ~     (p=0.739 n=10+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       689MB ± 2%     691MB ± 2%     ~     (p=0.579 n=10+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12     446MB ± 1%     445MB ± 1%     ~     (p=0.604 n=9+10)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                 15.8GB ± 1%    15.8GB ± 1%     ~     (p=0.796 n=10+10)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12       739MB ± 1%     742MB ± 1%     ~     (p=0.353 n=10+10)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12     491MB ± 2%     490MB ± 1%     ~     (p=0.853 n=10+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                 15.3GB ± 1%    15.3GB ± 1%     ~     (p=0.529 n=10+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       802MB ± 2%     799MB ± 1%     ~     (p=0.549 n=9+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     542MB ± 2%     546MB ± 3%     ~     (p=0.529 n=10+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                 14.6GB ± 1%    14.6GB ± 1%     ~     (p=0.529 n=10+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12       855MB ± 1%     861MB ± 2%     ~     (p=0.173 n=8+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12     591MB ± 3%     587MB ± 2%     ~     (p=0.353 n=10+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                 3.13GB ±17%    2.98GB ±19%     ~     (p=0.353 n=10+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12      10.9MB ±23%    11.0MB ±12%     ~     (p=0.720 n=10+9)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12    5.09MB ± 6%    5.10MB ± 4%     ~     (p=0.447 n=9+10)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                 3.10GB ±17%    2.76GB ± 3%     ~     (p=0.122 n=10+8)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12      11.6MB ±23%    11.2MB ± 8%     ~     (p=0.497 n=10+9)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12    5.50MB ± 6%    5.62MB ± 4%     ~     (p=0.218 n=10+10)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                 2.99GB ±20%    2.99GB ±18%     ~     (p=0.853 n=10+10)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12      12.0MB ± 8%    11.7MB ±16%     ~     (p=0.315 n=10+10)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12    5.83MB ± 3%    5.83MB ± 2%     ~     (p=0.968 n=9+10)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       633MB ± 5%     641MB ± 8%     ~     (p=0.549 n=9+10)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12     454MB ± 7%     442MB ± 5%     ~     (p=0.113 n=10+9)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12      13.2MB ±16%    12.9MB ± 4%     ~     (p=0.905 n=9+10)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12    5.83MB ± 5%    5.84MB ± 5%     ~     (p=0.968 n=9+10)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                 17.6GB ± 4%    17.6GB ± 5%     ~     (p=0.853 n=10+10)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       713MB ± 7%     731MB ± 5%     ~     (p=0.059 n=8+9)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     528MB ± 6%     527MB ± 5%     ~     (p=0.968 n=9+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                 3.27GB ±16%    3.41GB ± 8%     ~     (p=0.243 n=10+9)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12      12.3MB ± 3%    12.7MB ± 6%     ~     (p=0.182 n=9+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12    5.68MB ± 4%    5.71MB ± 3%     ~     (p=0.481 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                 16.8GB ± 0%    16.9GB ± 0%   +0.25%  (p=0.010 n=7+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12       588MB ± 2%     595MB ± 2%   +1.16%  (p=0.019 n=10+10)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                 2.80GB ± 8%    3.19GB ±19%  +14.18%  (p=0.006 n=9+10)

name                                                                                       old allocs/op  new allocs/op  delta
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                  95.3M ± 0%     95.4M ± 0%     ~     (p=0.280 n=10+10)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12       1.58M ± 8%     1.55M ± 2%     ~     (p=0.218 n=10+10)
SqlStatsPersisted/3node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12     1.21M ± 9%     1.18M ± 3%     ~     (p=0.353 n=10+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12       1.79M ± 5%     1.79M ± 4%     ~     (p=0.912 n=10+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12     1.34M ± 6%     1.33M ± 7%     ~     (p=0.579 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12       2.28M ± 4%     2.30M ± 5%     ~     (p=0.529 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12     1.77M ± 1%     1.81M ± 4%     ~     (p=0.122 n=8+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       2.78M ± 5%     2.77M ± 2%     ~     (p=0.796 n=10+10)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12     2.22M ± 1%     2.21M ± 1%     ~     (p=0.113 n=9+9)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12       3.40M ± 2%     3.44M ± 3%     ~     (p=0.063 n=10+10)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12     2.81M ± 1%     2.81M ± 1%     ~     (p=1.000 n=10+9)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       4.19M ± 3%     4.14M ± 2%     ~     (p=0.123 n=10+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     3.54M ± 3%     3.53M ± 1%     ~     (p=0.739 n=10+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                  87.5M ± 0%     87.6M ± 0%     ~     (p=0.203 n=8+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12       4.94M ± 2%     4.96M ± 2%     ~     (p=0.393 n=10+10)
SqlStatsPersisted/3node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12     4.27M ± 2%     4.27M ± 2%     ~     (p=0.912 n=10+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsFlush-12                  24.6M ±14%     23.4M ±17%     ~     (p=0.393 n=10+10)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectStatements-12       37.9k ± 3%     40.4k ± 8%     ~     (p=0.114 n=8+9)
SqlStatsPersisted/6node/computedColumns=6/BenchmarkPersistedSqlStatsSelectTransactions-12     28.2k ± 1%     28.3k ± 1%     ~     (p=0.203 n=8+10)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                  24.2M ±13%     21.7M ± 2%     ~     (p=0.573 n=10+8)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectStatements-12       45.0k ±21%     44.3k ±10%     ~     (p=0.971 n=10+10)
SqlStatsPersisted/6node/computedColumns=5/BenchmarkPersistedSqlStatsSelectTransactions-12     31.5k ± 0%     31.9k ± 3%     ~     (p=0.114 n=8+9)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                  23.1M ±19%     22.2M ± 5%     ~     (p=0.829 n=10+8)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectStatements-12       46.9k ± 1%     46.9k ± 1%     ~     (p=1.000 n=8+8)
SqlStatsPersisted/6node/computedColumns=4/BenchmarkPersistedSqlStatsSelectTransactions-12     36.2k ± 1%     36.2k ± 1%     ~     (p=0.863 n=9+9)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                  98.1M ± 4%     96.6M ± 7%     ~     (p=0.211 n=10+9)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectStatements-12       2.78M ± 2%     2.79M ± 3%     ~     (p=0.739 n=10+10)
SqlStatsPersisted/6node/computedColumns=3/BenchmarkPersistedSqlStatsSelectTransactions-12     2.23M ± 2%     2.25M ± 4%     ~     (p=0.529 n=10+10)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectStatements-12       51.8k ± 1%     51.9k ± 1%     ~     (p=0.438 n=8+9)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsSelectTransactions-12     36.5k ± 1%     36.8k ± 3%     ~     (p=0.497 n=9+10)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                  93.3M ± 7%     94.2M ± 6%     ~     (p=0.853 n=10+10)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectStatements-12       4.02M ± 2%     4.02M ± 3%     ~     (p=0.863 n=9+9)
SqlStatsPersisted/6node/computedColumns=1/BenchmarkPersistedSqlStatsSelectTransactions-12     3.36M ± 1%     3.36M ± 3%     ~     (p=0.780 n=9+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectStatements-12       50.6k ± 1%     50.7k ± 1%     ~     (p=0.779 n=8+10)
SqlStatsPersisted/3node/computedColumns=1/BenchmarkPersistedSqlStatsFlush-12                  89.4M ± 0%     89.5M ± 0%   +0.13%  (p=0.035 n=10+9)
SqlStatsPersisted/3node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                  90.4M ± 0%     90.6M ± 0%   +0.20%  (p=0.000 n=8+9)
SqlStatsPersisted/3node/computedColumns=3/BenchmarkPersistedSqlStatsFlush-12                  91.6M ± 0%     91.8M ± 0%   +0.23%  (p=0.004 n=9+10)
SqlStatsPersisted/3node/computedColumns=5/BenchmarkPersistedSqlStatsFlush-12                  95.3M ± 0%     95.6M ± 1%   +0.30%  (p=0.035 n=10+10)
SqlStatsPersisted/3node/computedColumns=4/BenchmarkPersistedSqlStatsFlush-12                  92.7M ± 0%     93.0M ± 1%   +0.33%  (p=0.006 n=8+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsSelectTransactions-12     36.8k ± 0%     36.9k ± 1%   +0.36%  (p=0.013 n=8+10)
SqlStatsPersisted/6node/computedColumns=0/BenchmarkPersistedSqlStatsFlush-12                  25.1M ±15%     27.1M ± 3%   +8.09%  (p=0.034 n=10+8)
SqlStatsPersisted/6node/computedColumns=2/BenchmarkPersistedSqlStatsFlush-12                  21.6M ± 5%     24.8M ±16%  +14.82%  (p=0.002 n=9+10)

```